feat(profiling-node): remove duplicate type declaration in nodejs profiling
